LayoutTests/platform/chromium/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt

CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "null" from accessing a frame at "http://127.0.0.1:8000".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: line 4: Uncaught TypeError: Cannot call method 'getElementById' of undefined
PASS
